refactor(hw_cluster): use f64 + scoped allow for the precision-loss casts
Tier 1 item 3. Two changes:

  - Bumped every f32 in hw_cluster/{scoring,pin_unpin,apply,mod,tests}.rs
    to f64. The scoring algorithm has zero perf reason to prefer f32,
    and f64's 52-bit mantissa (vs f32's 23) eliminates any plausible
    precision concern at our scale.
  - Renamed `print_table_f32_score` -> `print_score_table`; the `f32`
    suffix in the name was implementation detail leakage. Updated the
    six call sites in pin_unpin.rs and two `0f32` numeric literals.
  - `cast_precision_loss` still fires on f64 against 64-bit usize
    (52-bit mantissa < 64-bit pointer), but at our value range
    (component counts in the low thousands; 2^52 ≈ 4.5e15) it cannot
    realistically lose precision. Added a function-level
    `#[allow(clippy::cast_precision_loss)]` on the two scoring
    functions with a comment justifying the allow so the reasoning
    survives the next code review.

Verified: cargo build + fmt + clippy clean; precision_loss warnings
drop from 5 -> 0; all 358 workspace tests pass.

+++ crates/manta-server/src/service/hw_cluster/tests.rs
@@ -139,7 +139,7 @@ fn keep_iterating_mixed_components() {
 
 #[test]
 fn best_candidate_empty_inputs() {
-  let mut scores: Vec<(String, f32)> = vec![];
+  let mut scores: Vec<(String, f64)> = vec![];
   let hw: Vec<(String, HashMap<String, usize>)> = vec![];
   assert!(get_best_candidate_in_hsm(&mut scores, &hw).is_none());
 }
